* made plugin import implicitly * removed unnecessary code
