+++ .gitignore
@@ -17,3 +17,5 @@ coverage.xml
 .pytest_cache/
 .DS_Store
 *.sqlite3
+/tests/test_server/test_app/migrations/
+notes
