--- setup.cfg
@@ -1,4 +1,0 @@
-[metadata]
-url = https://github.com/Ori-Roza/drf-api-action
-author = Ori-Roza
-author_email = ori75660@gmail.com
